requirements.txt
annotated-types==0.5.0
anyio==4.13.0
asttokens==2.4.0
backcall==0.2.0
blinker==1.8.2
blis==0.7.11
Bootstrap-Flask==2.5.0
catalogue==2.0.10
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.1.4
confection==0.1.3
cvzone==1.6.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
executing==1.2.0
Flask==3.0.0
Flask-CKEditor==1.0.0
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
googletrans==4.0.0rc1
greenlet==3.3.0
groq==1.1.2
gTTS==2.5.4
gunicorn==24.1.1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
ipykernel==6.25.2
ipython==8.15.0
is-disposable-email==1.0.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
langcodes==3.3.0
langdetect==1.0.9
langid==1.1.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
murmurhash==1.0.10
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
parso==0.8.3
pathy==0.10.2
pdf_processor==0.2
pickleshare==0.7.5
pillow==12.1.1
platformdirs==3.10.0
preshed==3.0.9
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg==3.3.2
psycopg-binary==3.3.2
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyMuPDF==1.27.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
pyzmq==25.1.1
regex==2023.8.8
reportlab==4.4.10
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.4.8
stack-data==0.6.2
thinc==8.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typer==0.7.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.0.5
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.1
Werkzeug==3.0.4
WTForms==3.2.1
wtforms-validators==1.0.0
